Move journal non-integration tests.
